Move everything out of a fragment and into the main activity
