--- app/src/main/res/layout-land/fragment_gamepad.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-    <FrameLayout
-        android:layout_width="@dimen/config_gamepad_size"
-        android:layout_height="match_parent"
-        android:layout_marginStart="@dimen/gamepad_padding_horizontal"
-        android:layout_marginVertical="@dimen/gamepad_padding_vertical"
-        android:layout_alignParentStart="true"
-        android:layout_centerVertical="true"
-        android:id="@+id/left_container" />
-    <FrameLayout
-        android:layout_width="@dimen/config_gamepad_size"
-        android:layout_height="match_parent"
-        android:layout_marginEnd="@dimen/gamepad_padding_horizontal"
-        android:layout_marginVertical="@dimen/gamepad_padding_vertical"
-        android:layout_alignParentEnd="true"
-        android:layout_centerVertical="true"
-        android:id="@+id/right_container" />
-</RelativeLayout>
