--- app/src/main/res/layout/fragment_gamepad.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:orientation="horizontal"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:baselineAligned="false">
-    <FrameLayout
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_weight="0.5"
-        android:layout_marginStart="@dimen/gamepad_padding_horizontal"
-        android:layout_marginBottom="@dimen/gamepad_padding_vertical"
-        android:id="@+id/left_container" />
-    <FrameLayout
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_weight="0.5"
-        android:layout_marginEnd="@dimen/gamepad_padding_horizontal"
-        android:layout_marginBottom="@dimen/gamepad_padding_vertical"
-        android:id="@+id/right_container" />
-</LinearLayout>
